refactor(tensor): eliminate _tensor_helpers.py, move helpers into _C.py
Move _compute_strides, _bf16_to_f32, _f32_to_bf16 into _C.py as module-level
functions. Import _StrideTuple directly from Cython. Replace _HookHandle with
RemovableHandle from utils/hooks.py (matching torch's approach).
